+++ tests/requirements.py
@@ -21,10 +21,13 @@ class Requirements:
     DATUM_244 = "Add Snyk integration"
     DATUM_267 = "Add Image zip format"
     DATUM_280 = "Support KITTI dataset formats"
+    DATUM_283 = "Create cli tests for testing convert command for VOC format"
 
     # GitHub issues (bugs)
     # https://github.com/openvinotoolkit/datumaro/issues
     DATUM_BUG_219 = "Return format is not uniform"
+    DATUM_BUG_257 = "Dataset.filter doesn't count removed items"
+    DATUM_BUG_259 = "Dataset.filter fails on merged datasets"
 
 
 class SkipMessages: